Configure whether fd should be closed on drop
When the file descriptor is provided via the configuration, it is not
necessarily true that the file descriptor should be closed by the tun2
library, as it might be used outside of its scope.

To enable more flexibility, configure whether the closing should happen
in the library or be left as the responsability of the user.